[master] 7/7 Something special
Description:

It seems like our potential students are upset because of the last change in our algorithm. Instead of passing one exam, many of them need to worry about passing two exams with good scores. We are ready to give them a second chance. Now they can try to pass a special university admission exam! If the candidate's score on this exam is better than their mean score for the finals required by a Department, the university is ready to discard the results of the finals.

Objectives:

At this stage, your program should:

1. Read an N integer from the input. This integer represents the maximum number of students for each department.
2. Read the file named applicants.txt once again. Mind one additional column, right after the last exam's result. This column represents the special exam's score. For example, Willie McBride 76 45 79 80 100 Physics Engineering Mathematics(where 100 is the admission exam's score).
3. Choose the best score for a student in the ranking: either the mean score for the final exam(s) or the special exam's score. Use the same set of finals for each Department as in the previous stage. Note that you may need to compare the values several times: for example, if a student doesn't get accepted to the Department of the first priority, compare the finals mean score and the special exam's score once again (but this time, for the second priority department).
4. Keep the rest of the steps the same as in the previous stage. Once again, there should be no more than N accepted applicants for each department; use the same principles for sorting.
5. Output the names and the student's best score, either the mean finals score or the special exam's score to the files.